feat: decoupling data access and views
