+++ app/src/main/java/id/my/kaorikizuna/incu8tor/MainActivity.kt
@@ -7,9 +7,13 @@ import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.ui.platform.LocalContext
+import id.my.kaorikizuna.incu8tor.model.DeviceDetail
+import id.my.kaorikizuna.incu8tor.ui.addevice.AddDeviceScreen
 import id.my.kaorikizuna.incu8tor.ui.deviceConfiguration.DeviceConfigurationScreen
-import id.my.kaorikizuna.incu8tor.ui.deviceConfiguration.DeviceConfigurationScreenPreview
 import id.my.kaorikizuna.incu8tor.ui.home.HomeScreen
 import id.my.kaorikizuna.incu8tor.ui.theme.Incu8torTheme
 import id.my.kaorikizuna.incu8tor.viewmodel.DeviceViewModel
@@ -20,11 +24,29 @@ class MainActivity : ComponentActivity() {
         enableEdgeToEdge()
         setContent {
             Incu8torTheme {
-//                val viewModel = DeviceViewModel()
+                val viewModel = DeviceViewModel()
 //                val deviceDetail = viewModel.getDevice("24:DC:C3:45:EA:CC")
 //                Log.d(TAG, "device: ${deviceDetail.toString()}")
-                DeviceConfigurationScreenPreview()
+//                DeviceConfigurationScreenPreview()
 //                HomeScreen()
+
+                // testing purposes
+                val (deviceDetail, setDeviceDetail) = remember { mutableStateOf(DeviceDetail()) }
+                LaunchedEffect(Unit) {
+                    viewModel.getDevice("24:DC:C3:45:EA:CC", onSuccess = {
+                        setDeviceDetail(it)
+                    })
+                }
+                if (deviceDetail != DeviceDetail()) {
+                    DeviceConfigurationScreen(
+                        deviceDetail = deviceDetail,
+                        onUpdate = { deviceDetail -> viewModel.updateDevice(deviceDetail) },
+                        onDelete = { deviceDetail -> viewModel.deleteDevice(deviceDetail) })
+                }
+
+//                AddDeviceScreen(onSave = { deviceDetail ->
+//                    viewModel.addDevice(deviceDetail)
+//                })
             }
         }
     }
